Refactor HOCs’s folder structure
- Move AuthRoute, PrivateRoute, and ReferralRoute components from src/HOCs to src/HOCs/routes
- Move Loadable component from src/HOCs to src/HOCs/common
- Move toJS component from src/HOCs to src/HOCs/state
- Update paths to import statements
