+++ src/HOCs/common/Loadable.js
@@ -1,7 +1,7 @@
 // Module dependencies
 import LoadableVisibility from 'react-loadable-visibility/react-loadable';
 
-import Loader from '../components/common/Loader';
+import Loader from '../../components/common/Loader';
 
 // HOC
 const withLoadable = loader =>
