+++ src/HOCs/routes/AuthRoute.js
@@ -3,38 +3,27 @@ import React from 'react';
 import { connect } from 'react-redux';
 import { Redirect, Route } from 'react-router-dom';
 
-import { generateState } from '../helpers/data';
-import toJS from '../HOCs/toJS';
-import { getAuth } from '../data/session/reducers';
+import { generateState } from '../../helpers/data';
+import toJS from '../../HOCs/state/toJS';
+import { getAuth } from '../../data/session/reducers';
 
 // Constants
-import STATE_MODELS from '../constants/models/state';
-import PATHS from '../constants/router/paths';
-
-// Declare default props
-const defaultProps = {
-  exact: false
-};
+import STATE_MODELS from '../../constants/models/state';
+import PATHS from '../../constants/router/paths';
 
 // HOC
-const PrivateRoute = ({
-  component: Component,
-  state: { data: { authorization } },
-  exact,
-  ...rest
-}) => (
+const AuthRoute = ({ component: Component, state: { data: { authorization } }, ...rest }) => (
   <Route
     {...rest}
-    exact={exact}
     render={props => (
       <Choose>
-        <When condition={authorization}>
+        <When condition={!authorization}>
           <Component {...props} />
         </When>
         <Otherwise>
           <Redirect
             to={{
-              pathname: PATHS.users.signin,
+              pathname: PATHS.root,
               state: { from: props.location }
             }}
           />
@@ -44,15 +33,12 @@ const PrivateRoute = ({
   />
 );
 
-// Specify default values for props
-PrivateRoute.defaultProps = defaultProps;
-
 // Map state to props
 const mapStateToProps = state =>
   generateState(STATE_MODELS.immutable.setIn(['data', 'authorization'], getAuth(state)));
 
 // Connect component to application state
-const container = connect(mapStateToProps)(toJS(PrivateRoute));
+const container = connect(mapStateToProps)(toJS(AuthRoute));
 
 // Module exports
 export default container;
