+++ src/HOCs/routes/PrivateRoute.js
@@ -3,27 +3,38 @@ import React from 'react';
 import { connect } from 'react-redux';
 import { Redirect, Route } from 'react-router-dom';
 
-import { generateState } from '../helpers/data';
-import toJS from '../HOCs/toJS';
-import { getAuth } from '../data/session/reducers';
+import { generateState } from '../../helpers/data';
+import toJS from '../../HOCs/state/toJS';
+import { getAuth } from '../../data/session/reducers';
 
 // Constants
-import STATE_MODELS from '../constants/models/state';
-import PATHS from '../constants/router/paths';
+import STATE_MODELS from '../../constants/models/state';
+import PATHS from '../../constants/router/paths';
+
+// Declare default props
+const defaultProps = {
+  exact: false
+};
 
 // HOC
-const AuthRoute = ({ component: Component, state: { data: { authorization } }, ...rest }) => (
+const PrivateRoute = ({
+  component: Component,
+  state: { data: { authorization } },
+  exact,
+  ...rest
+}) => (
   <Route
     {...rest}
+    exact={exact}
     render={props => (
       <Choose>
-        <When condition={!authorization}>
+        <When condition={authorization}>
           <Component {...props} />
         </When>
         <Otherwise>
           <Redirect
             to={{
-              pathname: PATHS.root,
+              pathname: PATHS.users.signin,
               state: { from: props.location }
             }}
           />
@@ -33,12 +44,15 @@ const AuthRoute = ({ component: Component, state: { data: { authorization } }, .
   />
 );
 
+// Specify default values for props
+PrivateRoute.defaultProps = defaultProps;
+
 // Map state to props
 const mapStateToProps = state =>
   generateState(STATE_MODELS.immutable.setIn(['data', 'authorization'], getAuth(state)));
 
 // Connect component to application state
-const container = connect(mapStateToProps)(toJS(AuthRoute));
+const container = connect(mapStateToProps)(toJS(PrivateRoute));
 
 // Module exports
 export default container;
