+++ src/components/core/Navbar/index.js
@@ -8,7 +8,7 @@ import Dropdown, { DropdownContent, DropdownTrigger } from 'react-simple-dropdow
 import { action as toggleMenu } from 'redux-burger-menu/immutable';
 
 import { generateState } from '../../../helpers/data';
-import toJS from '../../../HOCs/toJS';
+import toJS from '../../../HOCs/state/toJS';
 import { signOut } from '../../../data/session/actions';
 import { getSession } from '../../../data/session/reducers';
 import { getAsync } from '../../../data/interfaces/session/reducers';
