+++ src/components/features/OAuth/Facebook/index.js
@@ -3,7 +3,7 @@ import { connect } from 'react-redux';
 import { bindActionCreators } from 'redux';
 
 import { generateState } from '../../../../helpers/data';
-import toJS from '../../../../HOCs/toJS';
+import toJS from '../../../../HOCs/state/toJS';
 import { oauthFacebook, oauthFailure, oauthRequest } from '../../../../data/session/actions';
 import { getSession } from '../../../../data/session/reducers';
