+++ src/screens/Users/screens/Profile/index.js
@@ -3,7 +3,7 @@ import { connect } from 'react-redux';
 import { bindActionCreators } from 'redux';
 
 import { generateState } from '../../../../helpers/data';
-import toJS from '../../../../HOCs/toJS';
+import toJS from '../../../../HOCs/state/toJS';
 import { deleteUser } from '../../../../data/session/actions';
 
 import * as creditsActions from '../../../../data/credits/actions';
