+++ src/screens/Users/screens/SignIn/index.js
@@ -4,7 +4,7 @@ import { Link } from 'react-router-dom';
 import { connect } from 'react-redux';
 
 import { generateState } from '../../../../helpers/data';
-import toJS from '../../../../HOCs/toJS';
+import toJS from '../../../../HOCs/state/toJS';
 import { getSession } from '../../../../data/session/reducers';
 import { getUI } from './reducers';
